refactor: Move profiler to memory_profiler at project root - Issue #680
- Moved src/profiler/ to memory_profiler/ at project root
- Updated include path in src/main.c to memory_profiler/memory_tracker.h
- Updated Makefile include paths
- Added -I. to Makefile CFLAGS for root-level includes
- All compilation tests pass

Closes #680
